Extracted utility classes
