+++ platformers/platformer_base.py
@@ -1,195 +1,14 @@
 import random
 from dataclasses import dataclass
-from enum import Enum
+from assets.enums import *
+from assets.configs import *
+from assets.sprites import *
+from assets.data import *
 
 import pygame
 from pygame import mixer
 
 from GameEnding import endscreen
-
-
-@dataclass
-class PlatformerConfig:
-    screen_width: int = 1000
-    screen_height: int = 1000
-    fps: int = 60
-    tile_size: int = 50
-    white: tuple = (255, 255, 255)
-    blue: tuple = (0, 0, 255)
-    world_data: list = None
-    font: pygame.font = pygame.font.SysFont('Bauhaus 93', 70)
-    font_score: pygame.font = pygame.font.SysFont('Bauhaus 93', 30)
-
-
-@dataclass
-class LanguageConfig:
-    max_levels: int
-
-
-class Colours(Enum):
-    WHITE = (255, 255, 255)
-    BLUE = (0, 0, 255)
-
-
-class Language(Enum):
-    ENGLISH = "English"
-    HINDI = "hindi"
-    BANGLA = "bangla"
-
-
-class LevelState(Enum):
-    START = 1
-    PLAYING = 2
-    PAUSED = 3
-    GAME_OVER = 4
-    EXIT = 5
-
-
-class PlatformerState(Enum):
-    MAIN_MENU = 1
-    PLAYING_LEVEL = 2
-    QUESTION = 3
-    QUIT = 4
-    WON = 5
-
-
-class QuestionState(Enum):
-    CORRECT = 1
-    INCORRECT = 2
-
-
-class CollisionType(Enum):
-    BLOB = 1
-    LAVA = 2
-    EXIT = 3
-    NONFATAL = 4
-    COIN = 5
-
-
-class PlayerEvent(Enum):
-    JUMP = 1
-    NOT_JUMPED = 2
-    LEFT = 3
-    RIGHT = 4
-    NONE = 5
-
-
-@dataclass
-class LanguageFx:
-    martyspeech_fx: pygame.mixer.Sound
-    coin_fx: pygame.mixer.Sound
-    jump_fx: pygame.mixer.Sound
-    game_over_fx: pygame.mixer.Sound
-
-
-class Button:
-    def __init__(self, x, y, image):
-        self.image = image
-        self.rect = self.image.get_rect()
-        self.rect.x = x
-        self.rect.y = y
-        self.clicked = False
-
-    def draw(self, screen):
-        action = False
-        # get mouse position
-        pos = pygame.mouse.get_pos()
-        # check mouseover and clicked conditions
-        if self.rect.collidepoint(pos):
-            if pygame.mouse.get_pressed()[0] == 1 and not self.clicked:
-                action = True
-                self.clicked = True
-        if pygame.mouse.get_pressed()[0] == 0:
-            self.clicked = False
-        # draw button
-        screen.blit(self.image, self.rect)
-        return action
-
-
-class Enemy(pygame.sprite.Sprite):
-    def __init__(self, x, y, image):
-        super().__init__()
-        self.image = image
-        self.rect = self.image.get_rect()
-        self.rect.x = x
-        self.rect.y = y
-        self.move_direction = 1
-        self.move_counter = 0
-
-    def update(self):
-        self.rect.x += self.move_direction
-        self.move_counter += 1
-        if abs(self.move_counter) > 50:
-            self.move_direction *= -1
-            self.move_counter *= -1
-
-
-class Platform(pygame.sprite.Sprite):
-    def __init__(self, x, y, move_x, move_y, image):
-        super().__init__()
-        self.image = pygame.transform.scale(image, (PlatformerConfig.tile_size, PlatformerConfig.tile_size // 2))
-        self.rect = self.image.get_rect()
-        self.rect.x = x
-        self.rect.y = y
-        self.move_counter = 0
-        self.move_direction = 1
-        self.move_x = move_x
-        self.move_y = move_y
-
-    def update(self):
-        self.rect.x += self.move_direction * self.move_x
-        self.rect.y += self.move_direction * self.move_y
-        self.move_counter += 1
-        if abs(self.move_counter) > 50:
-            self.move_direction *= -1
-            self.move_counter *= -1
-
-
-class Exit(pygame.sprite.Sprite):
-    def __init__(self, x, y, image):
-        pygame.sprite.Sprite.__init__(self)
-        self.image = pygame.transform.scale(image, (PlatformerConfig.tile_size, int(PlatformerConfig.tile_size * 1.5)))
-        self.rect = self.image.get_rect()
-        self.rect.x = x
-        self.rect.y = y
-
-
-class Coin(pygame.sprite.Sprite):
-    def __init__(self, x, y, image):
-        pygame.sprite.Sprite.__init__(self)
-        self.image = pygame.transform.scale(image, (PlatformerConfig.tile_size // 2, PlatformerConfig.tile_size // 2))
-        self.rect = self.image.get_rect()
-        self.rect.center = (x, y)
-
-
-class Lava(pygame.sprite.Sprite):
-    def __init__(self, x, y, image):
-        pygame.sprite.Sprite.__init__(self)
-        self.image = pygame.transform.scale(image, (PlatformerConfig.tile_size, PlatformerConfig.tile_size // 2))
-        self.rect = self.image.get_rect()
-        self.rect.x = x
-        self.rect.y = y
-
-
-@dataclass
-class Buttons:
-    start_button: Button
-    exit_button: Button
-    restart_button: Button
-
-
-@dataclass
-class LanguageImages:
-    sun_img: pygame.Surface
-    bg_img: pygame.Surface
-    restart_img: pygame.Surface
-    start_img: pygame.Surface
-    exit_img: pygame.Surface
-    martyspeech_img: pygame.Surface
-    coin_img: pygame.Surface
-    lava_img: pygame.Surface
-    exit: pygame.Surface
-    platform: pygame.Surface
 
 
 class World:
@@ -514,80 +333,4 @@ class Platformer:
         self.world.player.animate()
 
 
-class Player:
-    def __init__(self, x, y, langauge: Language):
-        self.language = langauge
-        self.images_right = []
-        self.images_left = []
-        self.index = 0
-        self.counter = 0
-        for num in range(1, 5):
-            img_right = pygame.image.load(f'{self.language.value}/img/guy{num}.png')
-            img_right = pygame.transform.scale(img_right, (40, 80))
-            img_left = pygame.transform.flip(img_right, True, False)
-            self.images_right.append(img_right)
-            self.images_left.append(img_left)
-        self.dead_image = pygame.image.load(f'{self.language.value}/img/ghost.png')
-        self.image = self.images_right[self.index]
-        self.rect = self.image.get_rect()
-        self.rect.x = x
-        self.rect.y = y
-        self.width = self.image.get_width()
-        self.height = self.image.get_height()
-        self.vel_y = 0
-        self.jumped = False
-        self.direction = 0
-        self.in_air = True
-        self.dx = 0
-        self.walk_cooldown = 5
-        self.col_thresh = 20
-        self.dy = 0
 
-    def reset(self, x, y):
-        self.__init__(x, y, self.language)
-
-    def reset_movement(self):
-        self.dx = 0
-        self.dy = 0
-
-    def reset_counter(self):
-        self.counter = 0
-        self.index = 0
-        if self.direction == 1:
-            self.image = self.images_right[self.index]
-        if self.direction == -1:
-            self.image = self.images_left[self.index]
-
-    def can_jump(self):
-        return not self.jumped and not self.in_air
-
-    def jump(self):
-        if self.can_jump():
-            self.vel_y = -15
-            self.jumped = True
-
-    def left(self):
-        self.dx -= 5
-        self.counter += 1
-        self.direction = -1
-
-    def right(self):
-        self.dx += 5
-        self.counter += 1
-        self.direction = 1
-
-    def animate(self):
-        if self.counter > self.walk_cooldown:
-            self.counter = 0
-            self.index += 1
-            if self.index >= len(self.images_right):
-                self.index = 0
-            if self.direction == 1:
-                self.image = self.images_right[self.index]
-            if self.direction == -1:
-                self.image = self.images_left[self.index]
-        self.vel_y += 1
-        if self.vel_y > 10:
-            self.vel_y = 10
-        self.dy += self.vel_y
-        self.in_air = True
